Reorganize alert files
