+++ src/app/core/alert/alert.component.html
@@ -9,7 +9,7 @@
       (click)="button.handler()"
       id="button.id"
       class="button"
-      [ngClass]="button.role === 'confirm' ? 'button-confirm' : 'button-cancel'">{{ button.text.toUpperCase() }}</button>
+      [ngClass]="button.role === 'confirm' ? 'button-confirm' : 'button-cancel'">{{ button.text }}</button>
   </footer>
 </div>
 </ng-container>
